+++ src/tests/tagLength.ts
@@ -3,6 +3,8 @@ import { CheerioAPI } from 'cheerio'
 function tagLength($: CheerioAPI, tagName = '*') {
     const tags = $(tagName)
 
+    console.log(tags)
+
     return tags.length
 }
