+++ src/utils/getDomainFromUrl.ts
@@ -1,0 +1,3 @@
+export function getDomainFromUrl(url: string) {
+    return new URL('/', url).origin
+}
