+++ src/utils/getPageReport.ts
@@ -4,6 +4,7 @@ import { IOptions } from '../types/options'
 import { Template, TestResult } from '../types/test'
 import { IReport } from '../types/report'
 import logger from './logger'
+import { didPass } from './didPass'
 
 export async function getPageReport(
     url: string,
@@ -41,16 +42,4 @@ export async function getPageReport(
         logger.error(`Tests failed: ${e}`)
         return { passed: false, url }
     }
-}
-
-export function didPass(status: number, tests?: Record<string, TestResult>) {
-    const haveAllTestsPassed = tests
-        ? Object.keys(tests).some((key) => !tests[key])
-        : true
-
-    if (status !== 200 || haveAllTestsPassed) {
-        return false
-    }
-
-    return true
 }
