--- src/utils/url.ts
@@ -1,5 +1,0 @@
-export function getDomainFromUrl(url: string) {
-    const urlClass = new URL('/', url)
-
-    return urlClass.origin
-}
